find-remotes

+++ conans/test/integration/metadata/test_metadata_deploy.py
@@ -1,4 +1,7 @@
+import os.path
 import textwrap
+
+import pytest
 
 from conans.test.utils.tools import TestClient
 
@@ -6,7 +9,8 @@ from conans.test.utils.tools import TestClient
 class TestMetadataDeploy:
     """ prove we can gather metadata too with a deployer"""
 
-    def test_deploy(self):
+    @pytest.fixture
+    def client(self):
         conanfile = textwrap.dedent("""
             import os
             from conan import ConanFile
@@ -30,17 +34,47 @@ class TestMetadataDeploy:
             def deploy(graph, output_folder, **kwargs):
                 conanfile = graph.root.conanfile
                 for r, d in conanfile.dependencies.items():
+                    if not os.path.exists(d.package_metadata_folder):
+                        continue
                     shutil.copytree(d.package_metadata_folder, os.path.join(output_folder, "pkgs",
                                                                             d.ref.name))
                     shutil.copytree(d.recipe_metadata_folder, os.path.join(output_folder, "recipes",
                                                                              d.ref.name))
            """)
 
-        c = TestClient()
+        c = TestClient(default_server_user=True, light=True)
         c.save({"conanfile.py": conanfile,
                 "deploy.py": deploy})
         c.run("create . --name=pkg1")
         c.run("create . --name=pkg2")
+        return c
+
+    def test_cache(self, client):
+        c = client
+        c.run("install --requires=pkg1/0.1 --requires=pkg2/0.1 --deployer=deploy")
+        assert "some logs pkg1!!!" in c.load("pkgs/pkg1/logs/mylogs.txt")
+        assert "some logs pkg2!!!" in c.load("pkgs/pkg2/logs/mylogs.txt")
+        assert "srclog pkg1!!!" in c.load("recipes/pkg1/logs/src.log")
+        assert "srclog pkg2!!!" in c.load("recipes/pkg2/logs/src.log")
+
+    def test_remote(self, client):
+        # But the remote story is more complex, metadata is not retrieved by default
+        c = client
+        c.run("upload * -c -r=default")
+        c.run("remove * -c")
+        # First install without metadata
+        c.run("install --requires=pkg1/0.1 --requires=pkg2/0.1")
+        # So this will not deploy metadata
+        c.run("install --requires=pkg1/0.1 --requires=pkg2/0.1 --deployer=deploy -f=json",
+              redirect_stdout="graph.json")
+        assert not os.path.exists(os.path.join(c.current_folder, "pkgs"))
+
+        # We can obtain the pkg-list for the graph, then "find-remote" and download the metadata
+        c.run("list -g=graph.json -f=json", redirect_stdout="mylist.json")
+        c.run("pkglist find-remote mylist.json -f=json", redirect_stdout="pkg_remotes.json")
+        c.run("download --list=pkg_remotes.json -r=default --metadata=*")
+
+        # Now we will have the metadata in cache and we can deploy it
         c.run("install --requires=pkg1/0.1 --requires=pkg2/0.1 --deployer=deploy")
         assert "some logs pkg1!!!" in c.load("pkgs/pkg1/logs/mylogs.txt")
         assert "some logs pkg2!!!" in c.load("pkgs/pkg2/logs/mylogs.txt")
